code/requirements.txt
torch==2.9.0
torchvision==0.21.0
numpy>=1.26
matplotlib>=3.8
Pillow>=10.0
tqdm>=4.66
